Upload Excel file containing depiction of DPS (#41)
Excel file containing interpretation of the DPS metdata. These standards are crafted to assist organizations in the initial stages of their data evaluation process, such as before conducting a privacy design review. They aim to enhance transparency during data exchanges within the organization. However, they are not meant to substitute for the operational data standards that govern day-to-day activities.